css-text/hyphens/hyphens-manual-003.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8">
<title>hyphens:manual, soft hyphens hidden by default</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://dev.w3.org/csswg/css-text-3/#word-break-property'>
<meta name="assert" content="When a word is not broken the UA will not display a hyphen-like glyph inside a word containing soft hyphens.">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style type='text/css'>
.test, .ref { border: 1px solid orange;  margin: 20px;  padding: 10px; width: 390px; font: 24px/1 csstest_asciiregular; white-space:nowrap; }
/* the CSS above is not part of the test */
</style>
</head>
<body>
<div id='log'></div>
<div class="test" lang="de" id="test"><span id="testspan">Eingabe&#x00AD;verarbeitungs&#x00AD;funktionen</span></div>
<div class="ref" lang="de" id="ref"><span id="refspan">Eingabeverarbeitungsfunktionen</div>
<script>
// make sure the box is large enough to hold the text without wrapping
var sentenceWidth = document.getElementById('testspan').offsetWidth
document.getElementById('test').style.width = String(sentenceWidth + 5)+'px'
document.getElementById('ref').style.width = String(sentenceWidth + 5)+'px'
// now run the test
test(function() {
assert_equals(document.getElementById('testspan').offsetWidth, document.getElementById('refspan').offsetWidth);
}, "words should be the same length");
</script>
</body>
</html>
